+++ eea/ldapadmin/import_export.py
@@ -45,6 +45,7 @@ def excel_headers_to_object(properties):
         'mobile': properties.get('mobile telephone number'),
         'fax': properties.get('fax number'),
         'organisation': properties.get('organisation'),
+        'department': properties.get('department'),
         'destinationIndicator': properties.get('reason to create')
     }