+++ eea/ldapadmin/nfp_nrc.py
@@ -603,14 +603,15 @@ class NfpNrc(SimpleItem, PropertyManager):
             form_data['user_id'] = user['uid']
 
         orgs = agent.all_organisations()
-        orgs = [{'id': k, 'text': v['name'], 'ldap': True}
-                for k, v in orgs.items()]
+        orgs = [{'id': k, 'text': v['name'], 'text_native': v['name_native'],
+                 'ldap': True} for k, v in orgs.items()]
 
         user_orgs = list(agent.user_organisations(user_id))
         if not user_orgs:
             org = form_data['organisation']
             if org:
-                orgs.append({'id': org, 'text': org, 'ldap': False})
+                orgs.append(
+                    {'id': org, 'text': org, 'text_native': '', 'ldap': False})
         else:
             org = user_orgs[0]
             org_id = agent._org_id(org)
@@ -620,7 +621,11 @@ class NfpNrc(SimpleItem, PropertyManager):
         choices = [('', '-')]
         for org in orgs:
             if org['ldap']:
-                label = u"%s (%s)" % (org['text'], org['id'])
+                if org['text_native']:
+                    label = u"%s (%s, %s)" % (org['text'], org['text_native'],
+                                              org['id'])
+                else:
+                    label = u"%s (%s)" % (org['text'], org['id'])
             else:
                 label = org['text']
             choices.append((org['id'], label))
@@ -925,16 +930,21 @@ class CreateUser(BrowserView):
         agent = self.context._get_ldap_agent()
         agent_orgs = self.orgs_in_country(nfp_country)
 
-        orgs = [{'id': k, 'text': v['name'], 'ldap':True}
-                for k, v in agent_orgs.items()]
+        orgs = [{'id': k, 'text': v['name'], 'text_native': v['name_native'],
+                 'ldap':True} for k, v in agent_orgs.items()]
         org = form_data.get('organisation')
         if org and not (org in agent_orgs):
-            orgs.append({'id': org, 'text': org, 'ldap': False})
+            orgs.append({'id': org, 'text': org, 'text_native': '',
+                         'ldap': False})
         orgs.sort(lambda x, y: cmp(x['text'], y['text']))
         choices = [('', '-')]
         for org in orgs:
             if org['ldap']:
-                label = u"%s (%s)" % (org['text'], org['id'])
+                if org['text_native']:
+                    label = u"%s (%s, %s)" % (org['text'], org['text_native'],
+                                              org['id'])
+                else:
+                    label = u"%s (%s)" % (org['text'], org['id'])
             else:
                 label = org['text']
             choices.append((org['id'], label))