+++ eea/ldapadmin/users_admin.py
@@ -483,16 +483,22 @@ class UsersAdmin(SimpleItem, PropertyManager):
         else:
             agent_orgs = self.orgs_in_country(nfp_country)
 
-        orgs = [{'id': k, 'text': v['name'], 'ldap':True}
+        orgs = [{'id': k, 'text': v['name'], 'text_native': v['name_native'],
+                 'ldap':True}
                 for k, v in agent_orgs.items()]
         org = form_data.get('organisation')
         if org and not (org in agent_orgs):
-            orgs.append({'id': org, 'text': org, 'ldap': False})
+            orgs.append({'id': org, 'text': org, 'text_native': '',
+                         'ldap': False})
         orgs.sort(lambda x, y: cmp(x['text'], y['text']))
         choices = [('-', '-')]
         for org in orgs:
             if org['ldap']:
-                label = u"%s (%s)" % (org['text'], org['id'])
+                if org['text_native']:
+                    label = u"%s (%s, %s)" % (org['text'], org['text_native'],
+                                              org['id'])
+                else:
+                    label = u"%s (%s)" % (org['text'], org['id'])
             else:
                 label = org['text']
             choices.append((org['id'], label))
@@ -599,13 +605,14 @@ class UsersAdmin(SimpleItem, PropertyManager):
             form_data = user
 
         orgs = agent.all_organisations()
-        orgs = [{'id': k, 'text': v['name'], 'ldap': True}
-                for k, v in orgs.items()]
+        orgs = [{'id': k, 'text': v['name'], 'text_native': v['name_native'],
+                 'ldap': True} for k, v in orgs.items()]
         user_orgs = list(agent.user_organisations(user_id))
         if not user_orgs:
             org = form_data.get('organisation')
             if org:
-                orgs.append({'id': org, 'text': org, 'ldap': False})
+                orgs.append(
+                    {'id': org, 'text': org, 'text_native': '', 'ldap': False})
         else:
             org = user_orgs[0]
             org_id = agent._org_id(org)
@@ -640,7 +647,11 @@ class UsersAdmin(SimpleItem, PropertyManager):
         choices = [('', '-')]
         for org in orgs:
             if org['ldap']:
-                label = u"%s (%s)" % (org['text'], org['id'])
+                if org['text_native']:
+                    label = u"%s (%s, %s)" % (org['text'], org['text_native'],
+                                              org['id'])
+                else:
+                    label = u"%s (%s)" % (org['text'], org['id'])
             else:
                 label = org['text']
             choices.append((org['id'], label))