+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 1.4.95 (unreleased)
 ------------------------
-* remove search helper from view and edit [dumitval]
+* add permission to Zope managers to view nrc administration [dumitval]
+* add Department column in nrc administration [dumitval]
 * add department to the LDAP fields [dumitval]
 * add name in native language and search helper to user attributes [dumitval]